Switch tutorial to its own story
