+++ .codeclimate.yml
@@ -1,0 +1,15 @@
+version: '2'
+checks:
+  method-complexity:
+    config:
+      threshold: 30
+plugins:
+  eslint:
+    enabled: true
+    channel: 'eslint-7'
+    config:
+      ignore_warnings: true
+exclude_patterns:
+  - '**/node_modules/'
+  - '**/*.test.js'
+  - '**/test/'
