+++ README.md
@@ -1,9 +1,10 @@
 ![Banner](assets/banner.png)
 
-[![Maintainability](https://api.codeclimate.com/v1/badges/a6122e76dcb42d834772/maintainability)](https://codeclimate.com/github/exoRift/react-fluent-mobile/maintainability)
-
-[![Storybook](https://img.shields.io/badge/TRY%20IT%20OUT-STORYBOOK%20-ff69b4?style=for-the-badge&logo=storybook)](https://exorift.github.io/react-fluent-mobile?path=/story/fluentselectionmixin--tutorial)
 [![NPM](https://img.shields.io/npm/v/react-fluent-mobile?style=for-the-badge)](https://www.npmjs.com/package/react-fluent-mobile)
+[![Storybook](https://img.shields.io/badge/TRY%20IT%20OUT-STORYBOOK-ff69b4?style=for-the-badge&logo=storybook)](https://exorift.github.io/react-fluent-mobile?path=/story/fluentselectionmixin--tutorial)
+[![Gitter](https://img.shields.io/badge/CHAT%20WITH%20US-GITTER-f68d42?style=for-the-badge&logo=gitter)](https://gitter.im/exoRift/react-fluent-mobile)
+
+[![Maintainability](https://api.codeclimate.com/v1/badges/a6122e76dcb42d834772/maintainability)](https://codeclimate.com/github/exoRift/react-fluent-mobile/maintainability)
 
 # *What's the problem with mobile browsers?*
 ### Mobile web browsers are an adaptation of the PC browsing experience for your mobile device. As a result, many features found on desktop browsers are sloppily implemented in ways that just aren't meant for phones, degrading your browsing experience.
@@ -52,7 +53,7 @@ function Component (props) {
 }
 ```
 
-## *Snappy, fluid, purposeful interactions*
+## *Snappy, fluid, deliberate interactions*
 *INSERT GIF*
 
 <br/>
