+++ src/stories/Coalescence.stories.jsx
@@ -5,7 +5,11 @@ import {
   FluentSelectionMixin
 } from '../'
 
+import banner from '../../assets/banner.png'
+
 import './styles/index.css'
+import './styles/ContextMixin.css'
+import './styles/SelectionMixin.css'
 
 export default {
   argTypes: {
@@ -38,16 +42,15 @@ export default {
   }
 }
 
-export const Default = (args) => (
+export const Playground = (args) => (
   <div>
-    <span id='span1'>This is some sample text!</span>
     <FluentContextMixin {...args}/>
     <FluentSelectionMixin {...args}/>
 
-    <span id='span2'>This is past the mixin</span>
+    <span>This is some sample text</span>
 
     <a href='/'>This is clickable</a>
-    <br/><br/><br/><br/><br/><br/>
-    <input id='test'/>
+
+    <img className='banner' alt='banner' src={banner}/>
   </div>
 )
