+++ src/stories/styles/ContextMixin.css
@@ -1,0 +1,3 @@
+.banner {
+  width: 20vw;
+}
