+++ src/stories/styles/index.css
@@ -21,6 +21,27 @@
   border-radius: 15px;
 }
 
+.storylink {
+  display: inline-block;
+  color: #189eec;
+  transform: translateX(-0.5rem);
+  cursor: pointer;
+  white-space: nowrap;
+}
+.storylink:hover {
+  color: #1579b3;
+}
+.storylink::before {
+  content: '\250f';
+  display: inline-block;
+  transform: translate(20%, -45%);
+}
+.storylink::after {
+  content: '\251b';
+  display: inline-block;
+  transform: translate(-15%, 40%);
+}
+
 @media (pointer: coarse) {
   .suggestion {
     display: none;
